refactor: move scratchpad completions to query-language/ and rename webview folder
Phase 6: Move src/documentdb/scratchpad/completions/ to
src/documentdb/query-language/scratchpad-completions/
- 8 files moved (4 source + 4 tdd)
- 6 import paths updated

Phase 7: Rename src/webviews/documentdbQuery/ to
src/webviews/query-language-support/
- 2 external import paths updated (CollectionView, QueryEditor)
- All internal relative imports unchanged

Three-way split now visible:
- query-language/shared/ — platform-neutral core
- query-language/scratchpad-completions/ — VS Code API specific
- webviews/query-language-support/ — Monaco specific

No functionality changes. Build, lint, 373 tests pass.
