+++ src/documentdb/query-language/scratchpad-completions/scratchpadContextDetector.ts
