+++ src/documentdb/query-language/scratchpad-completions/tdd/dynamicSchemaIntegration.test.ts
