+++ src/documentdb/query-language/scratchpad-completions/tdd/hoverAndEdgeCases.test.ts
