+++ src/documentdb/query-language/scratchpad-completions/tdd/scratchpadContextDetector.test.ts
