+++ src/webviews/query-language-support/completionStore.test.ts
