+++ src/webviews/query-language-support/completionStore.ts
