+++ src/webviews/query-language-support/completions/createCompletionItems.ts
