+++ src/webviews/query-language-support/completions/index.ts
