+++ src/webviews/query-language-support/completions/jsGlobals.ts
