+++ src/webviews/query-language-support/completions/mapCompletionItems.ts
