+++ src/webviews/query-language-support/completions/typeSuggestions.ts
