+++ src/webviews/query-language-support/documentdbQueryCompletionProvider.test.ts
