+++ src/webviews/query-language-support/documentdbQueryCompletionProvider.ts
