+++ src/webviews/query-language-support/documentdbQueryHoverProvider.ts
