+++ src/webviews/query-language-support/documentdbQueryValidator.ts
