+++ src/webviews/query-language-support/isCursorInsideString.test.ts
