+++ src/webviews/query-language-support/isCursorInsideString.ts
