+++ src/webviews/query-language-support/registerLanguage.ts
