+++ src/webviews/query-language-support/tdd/completionBehavior.test.ts
